% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Binning Continous Varaibles Based on Information value}
\usage{
bin(target, predictor, nbin = 5, unit = 1, min.node.pct = 0.05,
  p = 0.05)
}
\arguments{
\item{target}{the response variable}

\item{predictor}{the continous variable to bin}

\item{nbin}{numbers of binning}

\item{unit}{the basic interval of the continous variable}

\item{min.node.pct}{the smallest sample proportion of the bins}

\item{p}{not used yet}
}
\value{
a \code{bin} object

\code{cuts} the cut points
\code{IV} the information value
\code{ivtable} an iv-table given by \code{\link[woe]{woe}}
}
\description{
\code{bin} bins a continous variable by maximizing the infomation value.
}
