% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label.factor}
\alias{label.factor}
\alias{label.numeric}
\title{Transform a variable into factor}
\usage{
label.factor(x, levels = NULL)

label.numeric(x, cuts, na.level = NULL)
}
\arguments{
\item{x}{a numeric or a vector which can be coerced to be a factor}

\item{levels}{designated levels in the result factor}

\item{cuts}{cut points to segment a numeric, using a right-close interval}

\item{na.level}{should NA be an independent level? NULL for default behavior which inducts}
}
\description{
These functions turn a variable into a factor, with \code{NA} as an independent
level called 'Missing'.
}
\section{Functions}{
\itemize{
\item \code{label.numeric}: 
}}

