% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{woe}
\alias{woe}
\title{WoE Calculation}
\usage{
woe(df, y, x, cutp = NULL, good = NULL, na.exclude = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{y}{response variable}

\item{x}{predictor}

\item{cutp}{default \code{NULL}. If not provided, \code{x} will be treated as a factor. If \code{x} is a factor, this parameter will be ignored. Otherwise, \code{x} will be cut into a factor.}

\item{good}{the name of good. If not provided, the first levels of \code{y} will be treated as good}

\item{na.exclude}{whether NA values should be removed or kept as an independent level. FALSE by default.}
}
\description{
WoE Calculation
}
\examples{
df = data.frame(y = sample(c(0, 1), 100, replace = TRUE), x = runif(100))
woe(df, 'y', 'x', cutp = c(0, 0.3, 0.6, 1), good = '0')
}
